Apply alpha-eigenvalue research and benchmark scripts on top of upstream
This single squashed commit brings the alpha-eigenvalue / kinetics solver
work and ICSBEP benchmark tooling onto upstream/develop, while explicitly
excluding all FREYA / GEF / Spriggs / ALPHANSO fission-library work
(preserved separately on the freya-fork-seed branch).

Source: hand-merged from claude/upstream-sync, with FREYA additions stripped.

Major additions (vs upstream/develop):
  - Alpha-eigenvalue solver (src/eigenvalue.cpp, src/ifp.cpp, src/state_point.cpp)
  - Bloch-Airy UCN sampling (src/bloch_airy.cpp, include/openmc/bloch_airy.h)
  - User-guide chapters: alpha_eigenvalues, kinetics, gravity benchmarks
  - ICSBEP / COG benchmark tooling (icsbep-cog/, examples/, tools/)
  - setup.sh installation helper
  - 88 non-FREYA commits' worth of refactoring, doc, test, and example work

Original 154 commits and FREYA Phases 2-8 are preserved on:
  - develop-pre-sync-snapshot (old develop tip)
  - freya-fork-seed (= origin/claude/upstream-sync — recommended FREYA fork base)

The IFP-importance binned tally work (Phase 2-4, Godiva benchmark) will
land on top of this commit in a follow-up commit.

+++ src/bank.cpp
@@ -85,7 +85,7 @@ void sort_fission_bank()
   // bank for each parent particle id
   std::exclusive_scan(simulation::progeny_per_particle.begin(),
     simulation::progeny_per_particle.end(),
-    simulation::progeny_per_particle.begin(), 0);
+    simulation::progeny_per_particle.begin(), static_cast<int64_t>(0));
 
   // We need a scratch vector to make permutation of the fission bank into
   // sorted order easy. Under normal usage conditions, the fission bank is
